requirements.txt
amqp==5.0.5
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
drf-orjson-renderer==1.1.4
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
influxdb-client==1.14.0
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
kombu==5.0.2
line-protocol-parser==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
orjson==3.5.0
packaging==20.9
paho-mqtt==1.5.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
Rx==3.1.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
waitress==1.4.4
wcwidth==0.2.5
whitenoise==5.2.0
